[DEV] remove app backends
