--- backends/constants.py
@@ -1,7 +1,0 @@
-# -*- coding: utf8 -*-
-
-
-class Constants(object):
-    SUCCESS = 0000
-    SYSTEMERROR = 9999
-    TOKEN = "a905b993b5035122abe7be3a5c13ce2e55047981"
