--- backends/services/accountservice.py
@@ -1,4 +1,0 @@
-# -*- coding: utf8 -*-
-import logging
-
-logger = logging.getLogger("default")
